1. update code

+++ wiki/src/main/java/tss2/wiki/control/ContentController.java
@@ -10,6 +10,7 @@ import tss2.wiki.control.impl.ContentServiceImpl;
 import tss2.wiki.model.WikiOutline;
 import tss2.wiki.model.WikiRecord;
 import tss2.wiki.model.WikiSession;
+import tss2.wiki.model.WikiUser;
 import tss2.wiki.vo.WikiEntryVO;
 
 import javax.servlet.http.HttpServletRequest;
@@ -41,7 +42,7 @@ public class ContentController {
      * @param map
      * @return
      */
-    @RequestMapping(value = "/add", method = RequestMethod.POST, produces="application/json;charset=UTF-8")
+    @RequestMapping(value = "", method = RequestMethod.POST, produces="application/json;charset=UTF-8")
     public @ResponseBody ResultMessage addEntry(HttpServletRequest request, @RequestBody Map map) {
         String operation = map.get("operation").toString();
         // TODO session control
@@ -50,6 +51,11 @@ public class ContentController {
         WikiSession session = ss.checkUser(request);
         if (session == null) return new ResultMessage(1, "Authentication Failed");
         return cs.process(new WikiEntryVO(session, map));
+    }
+
+    @RequestMapping(value = "", method = RequestMethod.PUT, produces="application/json;charset=UTF-8")
+    public @ResponseBody ResultMessage putEntry(HttpServletRequest request, @RequestBody Map map) {
+        return addEntry(request, map);
     }
 
     /**
@@ -92,8 +98,17 @@ public class ContentController {
         return new WikiResult(1, wikiRecord);
     }
 
-    @RequestMapping(value = "/delete", method = RequestMethod.GET)
-    public @ResponseBody ResultMessage deleteEntry(@RequestParam(value = "title") String title) {
+    @RequestMapping(value = "", method = RequestMethod.DELETE)
+    public @ResponseBody ResultMessage deleteEntry(HttpServletRequest request, @RequestParam(value = "title") String title) {
+        SessionService ss = new SessionServiceimpl();
+        WikiSession session = ss.checkUser(request);
+        if (session == null) {
+            return new ResultMessage(1, "Authentication Failed");
+        }
+        WikiUser user = session.getUser();
+        if (user.getType() != WikiUser.USER_ADMIN) {
+            return new ResultMessage(1, "Authentication Failed");
+        }
         WikiRecord wikiRecord = new WikiRecord(title);
         wikiRecord.delete();
         return new ResultMessage(0);
